Simplified settings
All the settings are stored at the beginning of the program (after the
license), with a little description before them.

The management of settings should be improved: I want to destroy this
kind of section. "days", "hours" and "shift" should be stored in the
database, and it should be possible to edit them directly from the gui.
The location of the database should be asked from the gui too, adding
also a nice "recent files" menu.

In few words: actually there's no reason to have a settings file or a
part in the program where to save default settings; so I want to remove
them because I think that they are a workaround :)